Skyrim Batch Generator
SBG is a program that generates a text file containing user-chosen console commands for Skyrim that can then be treated as a batch file, telling the game to perform said commands within the file in succession.

Users can customize what's in these text files, allowing them to do a variety of things in the game - such as altering their level, skills, or items - quickly and conveniently without having to memorize the commands themselves.